+++ icoscp/const.py
@@ -3,10 +3,10 @@
 
 """    
     Define global (to icoscp) constants
-    mainly path variables and Url's used throught the package
-    usage: import icoscp.const as CPC
-    CPC.ST_PATH
-    
+    mainly path variables and Url's used throughout the package
+    Usage:
+    >>> import icoscp.const as CPC
+    >>> CPC.CP_META
 """
 
 
@@ -47,3 +47,5 @@ DOC_PYLIB = 'https://icos-carbon-portal.github.io/pylib/'
 DOC_M_AUTH = 'https://icos-carbon-portal.github.io/pylib/modules/#authentication'
 DOC_FAQ = 'https://icos-carbon-portal.github.io/pylib/faq/'
 DOC_FAQ_WARNINGS = 'https://icos-carbon-portal.github.io/pylib/faq/#how-do-i-suppress-warnings'
+# Metadata-schema concepts from Carbon Portal's ontologies.
+CP_META = 'http://meta.icos-cp.eu/ontologies/cpmeta/'