+++ icoscp/stilt/stiltobj.py
@@ -25,6 +25,7 @@ import json
 import xarray as xr
 import icoscp.const as CPC
 from icoscp.stilt import timefuncs as tf
+from icoscp.sparql import sparqls, runsparql
 from icoscp import __version__ as release_version
 ##############################################################################
 
@@ -64,6 +65,8 @@ class StiltStation():
         self.icos = None
         self.years = None
         self.geoinfo = None
+        self.dobjs_list = None               # Store a list of associated dobjs
+        self.dobjs_valid = False        # If True, dobjs sparql query already executed        
 
         self._set(st_dict)
 
@@ -86,7 +89,7 @@ class StiltStation():
 
     def __str__(self):
         # by default called when a an 'object' is printed
-        
+
         out = {'id:': self.id,
                'name:': self.name,
                'lat:': self.lat,
@@ -196,8 +199,10 @@ class StiltStation():
                 #Get response in json-format and read it in to a numpy array:
                 output=np.asarray(response.json())
 
-                #Convert numpy array with STILT results to a pandas dataframe:
-                df = pd.DataFrame(output[:,:], columns=eval(columns))
+                #Convert numpy array with STILT results to a pandas dataframe
+                cols = columns[1:-1].replace('"','').replace(' ','')
+                cols = list(cols.split(','))
+                df = pd.DataFrame(output[:,:], columns=cols)
 
                 #Replace 'null'-values with numpy NaN-values:
                 df = df.replace('null',np.NaN)
@@ -292,7 +297,11 @@ class StiltStation():
         #Concatenate xarrays on time axis:
         fp = xr.open_mfdataset(fp_files, combine='by_coords',
                                data_vars='minimal', coords='minimal',
-                               compat='override', parallel=True)
+                               compat='override', parallel=True,
+                               decode_cf=False)
+
+        # now check for CF compatibility
+        fp = xr.decode_cf(fp)
 
         #Format time attributes:
         fp.time.attrs["standard_name"] = "time"
@@ -315,7 +324,7 @@ class StiltStation():
         """
         Please do use this function with caution. Only very expirienced user
         should load raw data.
-    
+
         Parameters
         ----------
         start_date : STR
@@ -325,7 +334,7 @@ class StiltStation():
         cols : LIST[STR]
             A list of valid column names. You can retrieve the full list
             with _raw_clumn_names.
-    
+
         Returns
         -------
         columns : Pandas DataFrame
@@ -334,9 +343,9 @@ class StiltStation():
         #Convert date-strings to date objs:
         s_date = tf.parse(start_date).strftime('%Y-%m-%d')
         e_date = tf.parse(end_date).strftime('%Y-%m-%d')
-        
+
         # validate column names:
-            
+
         # make sure isodate is in the request
         if 'isodate' not in cols:
             cols.append('isodate')
@@ -344,25 +353,27 @@ class StiltStation():
         if len(columns) <= 1:
             return False
         # provide double quotes in the list
-        columns = json.dumps(columns) 
-    
+        columns = json.dumps(columns)
+
         # Check input parameters:
         if e_date < s_date:
             return False
-            
+
         # create http header and payload:
-            
-        headers = {'Content-Type': 'application/json', 'Accept-Charset': 'UTF-8'}        
+
+        headers = {'Content-Type': 'application/json', 'Accept-Charset': 'UTF-8'}
         data = '{"columns": '+ str(columns) + ',"fromDate": "'+s_date+'", "toDate": "'+e_date+'", "stationId": "'+self.id+'"}'
         response = requests.post(CPC.STILTRAW, headers=headers, data=data)
-        
+
         if response.status_code != 500:
 
             #Get response in json-format and read it in to a numpy array:
             output=np.asarray(response.json())
 
-            #Convert numpy array with STILT results to a pandas dataframe:
-            df = pd.DataFrame(output[:,:], columns=eval(columns))
+            #Convert numpy array with STILT results to a pandas dataframe
+            cols = columns[1:-1].replace('"','').replace(' ','')
+            cols = list(cols.split(','))
+            df = pd.DataFrame(output[:,:], columns=cols)
 
             #Replace 'null'-values with numpy NaN-values:
             df = df.replace('null',np.NaN)
@@ -376,16 +387,47 @@ class StiltStation():
             #Set 'date'-column as index:
             df.set_index(['date'],inplace=True)
 
-    
+
         # track data usage
         self.__portalUse('timeseries')
         #Return dataframe:
         return df
+<<<<<<< HEAD
     
-    #Function that checks the selection of columns that are to be
-    #returned with the STILT timeseries model output:
-    def __columns(self, cols):
+    def get_dobj_list(self):
+        """
+        If the stiltstation has a corresponding ICOS station
+        this function will return a dictionary filled with corresponding
+        dataobjects PID's. A sparql query is executed with ICOS Station id
+        and the sampling height.
 
+        Returns
+        -------
+        DICT
+            A dictionary with the following keys:
+            [dobj,hasNextVersion,spec,fileName,size,submTime,timeStart,timeEnd]
+
+        """
+        
+        if not self.valid:
+            return
+        
+        if not self.dobjs_valid and self.icos:
+            # add corresponding data object with observations
+            query = sparqls.dobj_for_samplingheight(\
+                            self.icos['stationId'], \
+                            self.icos['SamplingHeight'])
+            df = runsparql.RunSparql(query, 'pandas').run()
+            
+            self.dobjs_list = df.to_dict('records')
+            self.dobjs_valid = True        
+        
+        return self.dobjs_list
+        
+        
+    def __columns(self, cols):
+        #Function that checks the selection of columns that are to be
+        #returned with the STILT timeseries model output:
         if cols:
             # Convert user-specified columns to lower case.
             cols = cols.lower()
@@ -397,7 +439,7 @@ class StiltStation():
 
         #Check columns-input:
         if cols=='default':
-            columns = ('["isodate","co2.stilt","co2.bio","co2.fuel","co2.cement",'+ 
+            columns = ('["isodate","co2.stilt","co2.bio","co2.fuel","co2.cement",'+
                        '"co2.background"]')
 
         elif cols=='co2':
@@ -468,14 +510,14 @@ class StiltStation():
                 'internal': True}}}
         server = 'https://cpauth.icos-cp.eu/logs/portaluse'
         requests.post(server, json=counter)
-        
+
     def _raw_column_names(self):
         '''
         The STILT model calculates many different variables. We provide
         sensible groups for users, see the documentationfor .get_ts()
         This function returns an exhaustive list for all columns, approximately
         600, which can be used in conjunction get_raw()
-        
+
 
         Returns
         -------